fix(scope): code optimization
